Commit All

+++ EKanbanWeb/Models/MsPart.cs
@@ -12,7 +12,7 @@ namespace EKanbanWeb.Models
         public int PartId { get; set; }
         public string PartNo { get; set; }
         public string PartName { get; set; }
-        public float LotSize { get; set; }
+        public double LotSize { get; set; }
         public string Supplier { get; set; }
         public string Zone { get; set; }
         public string Remarks { get; set; }
